pyproject.toml
[project]
name = "langchain-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "deepagents>=0.4.11",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "fastmcp>=3.1.1",
    "langchain>=1.2.11",
    "langchain-anthropic>=1.3.4",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-neo4j>=0.7.0",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.0",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langgraph-cli[inmem]>=0.4.19",
    "langgraph-sdk>=0.3.11",
    "langgraph-swarm>=0.1.0",
    "langsmith>=0.7.16",
    "mcp[cli]>=1.26.0",
    "nemoguardrails>=0.21.0",
    "neo4j>=6.1.0",
    "numpy>=2.4.3",
    "openevals>=0.1.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "pytest>=9.0.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "streamlit>=1.56.0",
    "unstructured[md]>=0.21.5",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "jupyterlab>=4.5.5",
    "langsmith>=0.7.16",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
